SkincareBot/requirements.txt
streamlit==1.43.2
langchain==0.3.20
langchain-google-genai==2.1.0
dotenv-python==0.0.1